feat: 기본적인 carousel 구현
