+++ src/debounce.js
@@ -1,0 +1,12 @@
+const debounce = interval => {
+  let timer = null;
+
+  return callback => {
+    timer && clearTimeout(timer);
+    timer = setTimeout(() => {
+      callback();
+    }, interval);
+  };
+};
+
+export default debounce;
